A Geometric Approach to Solve Economic Models
An equilibrium in economics models can be often represented as a set of inter-temporal
dierence (or dierential) equations. These equations involve states and controls variables and
their solution are the so called policy functions. The general idea behind the method we propose
here is to view these inter-temporal equations as geometric restrictions (manifolds), so that time
shifted variables (in the present and/or in the past) are just others variables in a geometric space.
Each manifold (correspondent to each equation) can be obtained without using any notion of
time invariant policy function and it can be therefore computed separately from the others. For
this reason this process is fully parallelizable. Moreover, it is usually fast because often each
equation contains at least one variable that can be back out analytically. All manifolds are then
intersected through sequential binary operations, this process is computationally ecient since it
can be seen as a multi-steps univariate numerical optimization. The manifold that survives this
process is a geometric restriction between the states and one control. Finally, the correspondent
policy function can be unveiled imposing the fact that the variables in this surviving geometric
restriction must be connected thorugh a time-invariant function.